Refactor and optimize code across multiple modules
- Updated chunked upload logic in Azure backend to use `div_ceil` for chunk count calculation.
- Simplified object creation in GCS backend by using struct initialization syntax.
- Removed unnecessary type casting for part numbers in MinIO and S3 backends.
- Improved benchmark tests by eliminating redundant async calls and using more concise assertions.
- Enhanced chunking strategy enum with default variant and removed redundant default implementation.
- Refactored delta encoding logic to use `or_default` for hash table entries.
- Updated merge strategies to use default attributes and removed unnecessary default implementations.
- Improved object type parsing and testing for clarity and consistency.
- Streamlined reflog and refs handling with better error checking.
- Enhanced streaming index and pack writing logic for performance.
- Fixed various test cases to ensure consistency and clarity in commit references.

+++ crates/mediagit-storage/src/minio.rs
@@ -528,7 +528,7 @@ impl MinIOBackend {
                     .bucket(&bucket)
                     .key(&key)
                     .upload_id(&upload_id)
-                    .part_number(part_num as i32)
+                    .part_number(part_num)
                     .body(Bytes::from(chunk_data.clone()).into())
                     .send()
                     .await
@@ -572,7 +572,7 @@ impl MinIOBackend {
             .into_iter()
             .map(|(part_num, etag)| {
                 aws_sdk_s3::types::CompletedPart::builder()
-                    .part_number(part_num as i32)
+                    .part_number(part_num)
                     .e_tag(etag)
                     .build()
             })
